+++ src/maryasha/plugins/economy/finance/cards/info.py
@@ -20,6 +20,8 @@
 # LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 # OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 # SOFTWARE.
+from asyncio import gather
+
 from crescent import command
 
 from crescent import Plugin
@@ -28,7 +30,22 @@ from crescent import Context
 from crescent.ext import locales
 from crescent.ext import kebab
 
-from hikari import GatewayBot
+from hikari import Embed
+
+from flare import Row
+from flare import text_select
+from flare import MessageContext
+
+from .....helpers.tools import author
+from .....modules.economy import card_level_cost
+
+from .....modules.users import load
+from .....modules.users import new
+
+from .....helpers.emojis import E_C
+from .....helpers.emojis import E_CC
+
+from .....constants import EMBED_STD_COLOR
 
 
 plugin = Plugin()
@@ -43,5 +60,56 @@ DESCRIPTION = locales.LocaleMap('cardsInfo', ru=ru_LL, en_US=en_US_LL)
 @kebab.ify
 @command(description=DESCRIPTION)
 class CardsInfo:
+    TITLE = 'Информация'
+
+
+    async def main(self) -> None:
+        CARDS = self.user.cards
+
+        options = []
+
+        for card in CARDS.items():
+            options.append(card[0])
+
+
+        @text_select(
+            placeholder='Карты',
+            options=options
+        )
+        async def menu(ctx: MessageContext):
+            _embed = Embed(title=self.TITLE, color=EMBED_STD_COLOR)
+            author(ctx.member, _embed)
+
+            numbers = ctx.values[0]
+
+            CARD = CARDS[numbers]
+            NEXT_LEVEL = CARD.level + 1
+
+            cost = card_level_cost(NEXT_LEVEL)
+
+            _embed.description = \
+                f'{E_C} **Стоимость повышения уровня**:' \
+                f'`{cost}`$'
+
+            await ctx.respond(embed=_embed)
+
+
+        self.embed.description = f'{E_CC} Выберите Вашу карту'
+        components = await gather(Row(menu()))
+
+        return components
+
+
     async def callback(self, ctx: Context) -> None:
-        await ctx.respond('pong')
+        self.uid = str(ctx.user.id)
+
+        self.embed = Embed(title=self.TITLE, color=EMBED_STD_COLOR)
+        author(ctx.member, self.embed)
+
+        new(self.uid)
+        self.users = load()
+        self.user = self.users[self.uid]
+
+        components = await self.main()
+
+        await ctx.respond(embed=self.embed, components=components, ephemeral=True)